DROTH-58 Fix Bus stop expire childs and lose bus stop connection after move

+++ digiroad2-oracle/src/main/scala/fi/liikennevirasto/digiroad2/masstransitstop/TerminalBusStopStrategy.scala
@@ -90,10 +90,11 @@ class TerminalBusStopStrategy(typeId : Int, massTransitStopDao: MassTransitStopD
 
     optionalPosition.map(updatePosition(id, roadLink))
 
-    val children = MassTransitStopOperations.getTerminalMassTransitStopChildren(properties.toSeq)
-
-    massTransitStopDao.deleteChildren(id)
-    massTransitStopDao.insertChildren(id, children)
+    if(properties.exists(p => p.publicId == terminalChildrenPublicId)){
+      val children = MassTransitStopOperations.getTerminalMassTransitStopChildren(properties.toSeq)
+      massTransitStopDao.deleteChildren(id)
+      massTransitStopDao.insertChildren(id, children)
+    }
 
     enrichBusStop(fetchAsset(id))._1
   }
@@ -117,7 +118,7 @@ class TerminalBusStopStrategy(typeId : Int, massTransitStopDao: MassTransitStopD
     val topLeft = Point(position.x - meters, position.y - meters)
     val bottomRight = Point(position.x + meters, position.y + meters)
     val boundingBoxFilter = OracleDatabase.boundingBoxFilter(BoundingRectangle(topLeft, bottomRight), "a.geometry")
-    val filter = s"where a.asset_type_id = $typeId and (($boundingBoxFilter )or tbs.terminal_asset_id = $terminalId)"
+    val filter = s"where a.asset_type_id = $typeId and ((($boundingBoxFilter ) and (a.valid_to is null or a.valid_to > sysdate)) or tbs.terminal_asset_id = $terminalId)"
     massTransitStopDao.fetchPointAssets(massTransitStopDao.withFilter(filter))
       .filter(r => GeometryUtils.geometryLength(Seq(position, Point(r.lon, r.lat))) <= meters || r.terminalId.contains(terminalId))
   }
